pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project] 
name = "medical-assistant-by-RAG"
version = "0.1.0"
description = "Production-ready RAG API"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Diyorbek Tirkashov" }]

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.9",
  "pydantic-settings>=2.11.0",
  "numpy>=1.26",
  "requests>=2.32",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "matplotlib>=3.8",
  "pypdf>=4.0",
  "sentence-transformers>=3.0",
  "chromadb>=0.5",
  "langchain>=0.2",
  "langchain-community>=0.2",
  "langchain-huggingface>=0.1.0",
  "text-generation>=0.7.0",
  "langchain-xai>=0.1.0",
  "python-multipart>=0.0.20"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "httpx>=0.27",
  "black>=24.8",
  "isort>=5.13",
  "flake8>=7.1",
  "mypy>=1.11",
  "pytest-cov>=5.0",
  "pre-commit>=3.8"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rag_project*"]
